Fix crash on Androd 4.x ARM.
The reason for the crash is that when we scan the maps, the solist of the linker has changed, causing the dynamic library we are reading ELF information to be dlclose by the linker.

So we try to find the mutex lock inside the linker and hold this lock while scanning the maps, like the system's dl_iterate_phdr.